spark/requirements.txt
bson
pymongo
pyspark
kafka-python
pandas